define Priority for Transaction-Interceptors, handle Exception in Transactional right, no tests yet.
