+++ TODO.md
@@ -4,7 +4,7 @@
 * speedup by reusing scanned configuration
 * idea: ejb-deployment-mode, only accepts Alternatives
 * inject of EjbContext not in CDI-Beans allowed
-* don't evaluate ejb-transactionmanagement at pure cdi-bean
+* Transactional-Interceptor-rollbackOn, dontRollbackOn
 
 # doing
 * XmlLessPersistenceFactory ease usage!!!
